refactor: split server startup modules
