+++ internal/app/modules/mcp/mcp.go
@@ -1,0 +1,18 @@
+package mcpmodule
+
+import (
+	"log"
+
+	"github.com/takutakahashi/agentapi-proxy/internal/app"
+	mcpiface "github.com/takutakahashi/agentapi-proxy/internal/interfaces/mcp"
+)
+
+// RegisterHandler registers the MCP HTTP handler.
+func RegisterHandler(proxyServer *app.Server) {
+	log.Printf("[MCP_HANDLER] Registering MCP handler...")
+
+	mcpHandler := mcpiface.NewMCPHandler(proxyServer)
+	proxyServer.AddCustomHandler(mcpHandler)
+
+	log.Printf("[MCP_HANDLER] MCP handler registered successfully at /mcp")
+}
